add metric
