+++ jsonrpc/metrics/metrics.go
@@ -8,20 +8,13 @@ import (
 )
 
 const (
-	prefix                    = "jsonrpc_"
-	requestPrefix             = prefix + "request_"
-	requestsHandledName       = requestPrefix + "handled"
-	requestDurationName       = requestPrefix + "duration"
-	connName                  = requestPrefix + "connection"
-	requestMethodName         = requestPrefix + "method"
-	requestMethodDurationName = requestPrefix + "method_duration"
-
-	wsRequestPrefix             = prefix + "ws_request_"
-	requestWsMethodName         = wsRequestPrefix + "method"
-	requestWsMethodDurationName = wsRequestPrefix + "method_duration"
+	prefix              = "jsonrpc_"
+	requestPrefix       = prefix + "request_"
+	requestsHandledName = requestPrefix + "handled"
+	requestDurationName = requestPrefix + "duration"
+	connName            = requestPrefix + "connection"
 
 	requestHandledTypeLabelName = "type"
-	requestMethodLabelName      = "method"
 )
 
 // RequestHandledLabel represents the possible values for the
@@ -51,9 +44,8 @@ const (
 // Register the metrics for the jsonrpc package.
 func Register() {
 	var (
-		counterVecs   []metrics.CounterVecOpts
-		histograms    []prometheus.HistogramOpts
-		histogramVecs []metrics.HistogramVecOpts
+		counterVecs []metrics.CounterVecOpts
+		histograms  []prometheus.HistogramOpts
 	)
 
 	counterVecs = []metrics.CounterVecOpts{
@@ -91,52 +83,9 @@ func Register() {
 		},
 	}
 
-	histogramVecs = []metrics.HistogramVecOpts{
-		{
-			HistogramOpts: prometheus.HistogramOpts{
-				Name:    requestMethodDurationName,
-				Help:    "[JSONRPC] Histogram for the runtime of requests",
-				Buckets: prometheus.LinearBuckets(start, width, count),
-			},
-			Labels: []string{requestMethodLabelName},
-		},
-		{
-			HistogramOpts: prometheus.HistogramOpts{
-				Name:    requestWsMethodDurationName,
-				Help:    "[JSONRPC] Histogram for the runtime of ws requests",
-				Buckets: prometheus.LinearBuckets(start, width, count),
-			},
-			Labels: []string{requestMethodLabelName},
-		},
-	}
-
 	metrics.RegisterCounterVecs(counterVecs...)
 	metrics.RegisterHistograms(histograms...)
 	metrics.RegisterHistogramVecs(histogramVecs...)
-}
-
-// WsRequestMethodDuration observes (histogram) the duration of a ws request from the
-// provided starting time.
-func WsRequestMethodDuration(method string, start time.Time) {
-	metrics.HistogramVecObserve(requestMethodDurationName, method, time.Since(start).Seconds())
-}
-
-// WsRequestMethodCount increments the ws requests handled counter vector by one for
-// the given method.
-func WsRequestMethodCount(method string) {
-	metrics.CounterVecInc(requestMethodName, method)
-}
-
-// RequestMethodDuration observes (histogram) the duration of a request from the
-// provided starting time.
-func RequestMethodDuration(method string, start time.Time) {
-	metrics.HistogramVecObserve(requestMethodDurationName, method, time.Since(start).Seconds())
-}
-
-// RequestMethodCount increments the requests handled counter vector by one for
-// the given method.
-func RequestMethodCount(method string) {
-	metrics.CounterVecInc(requestMethodName, method)
 }
 
 // CountConn increments the connection counter vector by one for the
