+++ jsonrpc/metrics/metrics_x1.go
@@ -1,0 +1,63 @@
+package metrics
+
+import (
+	"github.com/0xPolygonHermez/zkevm-node/metrics"
+	"github.com/prometheus/client_golang/prometheus"
+	"time"
+)
+
+var (
+	requestMethodName         = requestPrefix + "method"
+	requestMethodDurationName = requestPrefix + "method_duration"
+
+	wsRequestPrefix             = prefix + "ws_request_"
+	requestWsMethodName         = wsRequestPrefix + "method"
+	requestWsMethodDurationName = wsRequestPrefix + "method_duration"
+	requestMethodLabelName      = "method"
+
+	start         = 0.1
+	width         = 0.1
+	count         = 10
+	histogramVecs = []metrics.HistogramVecOpts{
+		{
+			HistogramOpts: prometheus.HistogramOpts{
+				Name:    requestMethodDurationName,
+				Help:    "[JSONRPC] Histogram for the runtime of requests",
+				Buckets: prometheus.LinearBuckets(start, width, count),
+			},
+			Labels: []string{requestMethodLabelName},
+		},
+		{
+			HistogramOpts: prometheus.HistogramOpts{
+				Name:    requestWsMethodDurationName,
+				Help:    "[JSONRPC] Histogram for the runtime of ws requests",
+				Buckets: prometheus.LinearBuckets(start, width, count),
+			},
+			Labels: []string{requestMethodLabelName},
+		},
+	}
+)
+
+// WsRequestMethodDuration observes (histogram) the duration of a ws request from the
+// provided starting time.
+func WsRequestMethodDuration(method string, start time.Time) {
+	metrics.HistogramVecObserve(requestMethodDurationName, method, time.Since(start).Seconds())
+}
+
+// WsRequestMethodCount increments the ws requests handled counter vector by one for
+// the given method.
+func WsRequestMethodCount(method string) {
+	metrics.CounterVecInc(requestMethodName, method)
+}
+
+// RequestMethodDuration observes (histogram) the duration of a request from the
+// provided starting time.
+func RequestMethodDuration(method string, start time.Time) {
+	metrics.HistogramVecObserve(requestMethodDurationName, method, time.Since(start).Seconds())
+}
+
+// RequestMethodCount increments the requests handled counter vector by one for
+// the given method.
+func RequestMethodCount(method string) {
+	metrics.CounterVecInc(requestMethodName, method)
+}
